+++ R/oncomatrix.R
@@ -330,16 +330,17 @@ bubble_plot = function(plot_dat, lab_dat = NULL, x_var = NULL, y_var = NULL,
   if(!is.null(lab_dat)){
     # points(x = lab_dat$x, y = lab_dat$y, cex = lab_dat$size_z,
     #        pch = 16, col = lab_dat$color_var)
-    if(nrow(lab_dat) > 0 & nrow(lab_dat) < 2){
-      text(x = lab_dat$x, y = lab_dat$y, labels = lab_dat$z_text, adj = 1, offset = 0.2, cex = text_size, col = lab_dat$color_var)
-    }else if(nrow(lab_dat) >= 2){
-      symbols(x = lab_dat$x, y = lab_dat$y, circles = lab_dat$size_z,
-              bg = lab_dat$color_var, add = TRUE, fg = "white", inches = 0.1)
-
-      wordcloud::textplot(x = lab_dat$x, y = lab_dat$y, words = lab_dat$z_text,
-                          cex = text_size, new = FALSE, show.lines = TRUE,
-                          xlim = x_lims[c(1, 4)], ylim = y_lims[c(1, 4)], font = 3, col = lab_dat$color_var)
-    }
+    text(x = lab_dat$x, y = lab_dat$y, labels = lab_dat$z_text, adj = 1, offset = 0.2, cex = text_size, col = lab_dat$color_var, xpd = TRUE)
+    # if(nrow(lab_dat) > 0 & nrow(lab_dat) < 2){
+    #   text(x = lab_dat$x, y = lab_dat$y, labels = lab_dat$z_text, adj = 1, offset = 0.2, cex = text_size, col = lab_dat$color_var)
+    # }else if(nrow(lab_dat) >= 2){
+    #   symbols(x = lab_dat$x, y = lab_dat$y, circles = lab_dat$size_z,
+    #           bg = lab_dat$color_var, add = TRUE, fg = "white", inches = 0.1)
+    #
+    #   wordcloud::textplot(x = lab_dat$x, y = lab_dat$y, words = lab_dat$z_text,
+    #                       cex = text_size, new = FALSE, show.lines = TRUE,
+    #                       xlim = x_lims[c(1, 4)], ylim = y_lims[c(1, 4)], font = 3, col = lab_dat$color_var)
+    # }
   }
 }